# docs/specs/openapi.yaml
openapi: "3.1.0"
info:
  title: Aether Platform API
  description: |
    White-label SaaS коммуникационная платформа.
    Multi-tenant, plugin architecture, AI-powered.

    ## Аутентификация
    - **Bearer JWT**: `Authorization: Bearer <access_token>`
    - **API Key**: `X-API-Key: ak_...`

    ## Tenant isolation
    Все tenant-specific эндпоинты требуют JWT с `organisation_id` claim.
    Admin эндпоинты требуют `is_superadmin = true`.

    ## Rate Limiting
    - Auth: IP-based (5 login/min, 3 signup/hour)
    - Tenant: per-plan (60-6000 RPM)
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    ## Pagination
    Cursor-based: `?cursor=<last_id>&limit=<1-100>`, default 20.
    Response: `{"items": [...], "next_cursor": "uuid"}`.

    ## Errors
    Все ошибки возвращают JSON: `{"error": {"code": "...", "message": "...", "details": {...}}}`
  version: "1.0.0"
  contact:
    name: Aether Platform
    email: dev@aether.cloud
servers:
  - url: https://api.aether.cloud/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
security:
  - bearerAuth: []
  - apiKeyAuth: []

# ============================================================================
# SECURITY SCHEMES
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/refresh
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API key from /auth/api-keys (prefix: ak_)'

  # ============================================================================
  # COMMON SCHEMAS
  # ============================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "invalid_credentials"
            message:
              type: string
              description: Human-readable error message (localized)
              example: "Неверный email или пароль"
            details:
              type: object
              description: Field-level validation errors
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Ошибка валидации"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # --- Pagination ---
    PaginatedResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {}
        next_cursor:
          type: string
          format: uuid
          nullable: true
          description: Cursor for next page, null if last page
        total:
          type: integer
          description: Total count (when available)
          example: 42

    # --- Auth ---
...
    ChannelCreate:
      type: object
      required: [channel_type, display_name]
      properties:
        channel_type:
          type: string
          enum: [telegram, web_widget, email, whatsapp, rest_api]
        display_name:
          type: string
          minLength: 1
          maxLength: 255
        config:
          type: object
          description: Channel-specific config (bot_token, smtp_host, etc.)
    ChannelResponse:
      type: object
      required: [id, channel_type, display_name, is_active, created_at]
      properties:
        id:
          type: string
          format: uuid
        channel_type:
          type: string
        display_name:
          type: string
        is_active:
          type: boolean
        priority:
          type: integer
        config:
          type: object
          description: Non-sensitive config (токены не возвращаются)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChannelTestResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
          description: Результат теста
          example: "Telegram bot token валиден, бот @my_bot активен"
        latency_ms:
          type: integer
          description: Задержка соединения

    # --- Conversations ---
...
  parameters:
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
        format: uuid
      description: Cursor для пагинации (id последнего элемента предыдущей страницы)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Количество элементов на странице

  # ============================================================================
  # COMMON RESPONSES
  # ============================================================================
  responses:
    401Unauthorized:
      description: Не авторизован (JWT истёк или отсутствует)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthorized"
              message: "Требуется аутентификация"
...
    404NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    422ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      summary: Список каналов
      tags: [Channels]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список каналов
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
    post:
      operationId: createChannel
      summary: Подключить канал
      tags: [Channels]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelCreate'
      responses:
        '201':
          description: Канал создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /channels/{channel_id}:
    get:
      operationId: getChannel
      summary: Детали канала
      tags: [Channels]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: channel_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали канала
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
    patch:
      operationId: updateChannel
      summary: Обновить канал
      tags: [Channels]
      security:
        - bearerAuth: []
      parameters:
        - name: channel_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 255
                is_active:
                  type: boolean
                config:
                  type: object
      responses:
        '200':
          description: Канал обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
    delete:
      operationId: deleteChannel
      summary: Удалить канал
      tags: [Channels]
      security:
        - bearerAuth: []
      parameters:
        - name: channel_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Канал удалён
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
  /channels/{channel_id}/test:
    post:
      operationId: testChannel
      summary: Протестировать подключение канала
      tags: [Channels]
      security:
        - bearerAuth: []
      parameters:
        - name: channel_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Результат теста
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelTestResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'

  # ==========================================================================
  # CONVERSATIONS
  # ==========================================================================
...